requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1